requirements.txt
# Core API dependencies
fastapi==0.94.1
uvicorn==0.22.0
pydantic==1.10.11
starlette==0.26.1

# Web framework and templating
Flask==2.3.2
Jinja2==3.1.2
Werkzeug==2.3.7
itsdangerous==2.1.2
blinker==1.6.2

# Database
pymongo==4.4.1
dnspython==2.4.2  # MongoDB SRV support

# HTTP and networking
requests==2.31.0
urllib3==2.0.4
h11==0.14.0
anyio==3.7.1
sniffio==1.3.0
idna==3.4
certifi==2023.7.22
charset-normalizer==3.2.0

# Configuration and environment
python-dotenv==1.0.0

# Testing
pytest==6.2.5
atomicwrites==1.4.1
attrs==23.1.0
iniconfig==2.0.0
pluggy==1.2.0
py==1.11.0
toml==0.10.2
colorama==0.4.6  # For colored terminal output in tests
exceptiongroup==1.1.3

# Type checking
typing_extensions==4.7.1

# Packaging and deployment
click==8.1.7
packaging==23.1
MarkupSafe==2.1.3